🐛 Remove support for rc file in the repository
There's a bug where `.current-thing.json` is not found if specified in
the repository directory. This is because the plugin is actually run
from the Xbar plugins directory, so the "current" directory is not the
one with the config file.

This doesn't actually fix the bug, it just removes that as a supported
configuration option.